requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
openai
langchain
langchain-openai
langchain-community
faiss-cpu
tqdm
pillow
pytesseract
PyMuPDF
unstructured